requirements.txt
Django==3.2.13
djangorestframework==3.12.0
djangorestframework-simplejwt==5.1.0
Pillow==9.1.0
requests==2.27.1